Better bash scripts and better data loss protection and additonal variables

+++ srcs/tools/env_delete.sh
@@ -1,6 +1,8 @@
-read -p "env_delete.sh: WARNING DANGER ZONE: Deleting your env file? (yes/n): " RESPONSE
-if [ "$RESPONSE" = 'yes' ]; then
-  rm -f ../.env
-else
-  echo "env_delete.sh: Aborting."
+if [ -s ../.env ]; then
+  read -p "env_delete.sh: WARNING DANGER ZONE: Deleting your env file? (yes/n): " RESPONSE
+  if [ "$RESPONSE" = 'yes' ]; then
+    rm -f ../.env
+  else
+    echo "env_delete.sh: Aborting."
+  fi
 fi
